BIGINT in mySql correspond to Long in java and it can be auto-increment column

+++ water-voucher/src/main/java/com/wv/controllers/CustomerControlller.java
@@ -34,7 +34,7 @@ public class CustomerControlller {
 	}
 	
 	@GetMapping("employees/{id}")
-	public Customer get(@PathVariable Integer id) {
+	public Customer get(@PathVariable Long id) {
 		return customerRepoJooqImpl.get(id).orElse(null);
 	}
 }
